Support underlying put API
It is common practice to pre-allocate arrays so in high write situations you don't need to keep 
creating large array objects. As the current API requires full array, this means an array needs to be 
copied before passing to RocksDB. 

By supporting the underlying C API, an additional memory copy can be avoided leading to
a big saving in GC in high write circumstances.